+++ internal/pkg/service/screenChangeHandler.go
@@ -1,0 +1,7 @@
+package service
+
+func ScreenChangeHandler() {
+	//pixel check
+	//screenSizeX, screenSizeY := robotgo.GetScreenSize()
+	//lastPixelColor := robotgo.GetPixelColor(int(screenSizeX/2), int(screenSizeY/2))
+}
